refactor: your vote pane to ProposalAccountVotingPane
