+++ packages/desktop/views/dashboard/governance/components/proposal-details/index.ts
@@ -1,1 +1,2 @@
+export { default as ProposalAccountVotingPane } from './ProposalAccountVotingPane.svelte'
 export { default as ProposalDetailsPane } from './ProposalDetailsPane.svelte'
